backend/requirements.txt
openai
python-dotenv
openai-agents
fastapi[standard]
fastmcp
requests
uvicorn
sqlmodel
python-jose[cryptography]
passlib[bcrypt]